requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.0
black==25.11.0
certifi==2026.2.25
cffi==2.0.0
click==8.1.8
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
idna==3.11
motor==3.7.1
mypy_extensions==1.1.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.4.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyJWT==2.12.1
pymongo==4.16.0
python-dotenv==1.2.1
pytokens==0.4.1
ruff==0.15.9
six==1.17.0
starlette==0.49.3
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
